Un-deprecate use of HttpRequest/Response interceptor
The browsermob-proxy RequestInterceptor and ResponseInterceptor don't
let you modify the response entity before it is returned to the browser,
among other things.

The HttpRequestInterceptor and HttpResponseInterceptor allow access to
the request after the proxying code has modified it, and to the response
before the proxying code has a chance to access it, giving callers a
chance to do many of the finer request / response modifications
requested.