Fix HarResultsTest to be right on all platforms
Test is designed to make sure that we're recording in the HAR the size
of the data transferred between proxy and web server, not proxy and
client.

In most cases, we can hard-code the expected values as they don't
change. However, the response headers in this test change depending on
which platform they're running on. So, for the response headers, we
make a direct connection from client to server, and use the direct
response headers to calculate what we should see when we run the test on
this platform.